perf(app): move book state and projection to a dedicated thread
The UI thread never touches the order-book history again: a BookEngine
(the old OrderflowView state machine, unchanged semantics and logging)
now lives on a quantick-book worker thread. The UI forwards feed
events and coalesced latest-wins projection requests through a channel
and only draws the newest published frame, so a dense-book projection
can delay a frame swap but never block a frame.

- orderflow_engine.rs: pure state machine + published snapshots; keeps
  the last built frame across soft invalidations so the heatmap never
  flashes empty between rebuilds.
- orderflow_worker.rs: command loop, latest-wins projection, test
  barrier (Flush) so the async pipeline stays deterministic in tests.
- orderflow_view.rs: UI facade; owns config mirror and grouping-change
  staging; adopts engine auto-base buckets published from live data.
