+++ crates/app/src/main.rs
@@ -18,8 +18,10 @@ mod config;
 mod feed;
 mod metrics;
 mod orderflow;
+mod orderflow_engine;
 mod orderflow_render;
 mod orderflow_view;
+mod orderflow_worker;
 mod price_view;
 mod state;
 mod style;
